requirements.txt
Flask==2.3.3
flask-restx==1.1.0
python-dotenv==1.0.0
openai==1.8.0
g